// src/service/type.ts
export enum CHART_TYPE {
    PROFIT = 'profit', 
    USER_ACQUISITION = 'user-acquisition', 
    DATA_USAGE = 'data-usage', 
}
export const MONTHS = [
    'Jan','Feb','Mar','Apr','May','Jun','July','Aug','Sep','Oct','Nov','Dec'
]
export enum FILTER_TYPE {
    THIS_YEAR = 'year', 
    THIS_MONTH = 'month', 
}